requirements.txt
streamlit
streamlit-authenticator==0.3.3
pandas
streamlit-ace
lightning
umap-learn
rdkit
plotly
autogluon
